Remove `ProspectiveParachainsMode` usage in backing subsystem (#6215)
Since async backing parameters runtime api is released on all networks
the code in backing subsystem can be simplified by removing the usages
of `ProspectiveParachainsMode` and keeping only the branches of the code
under `ProspectiveParachainsMode::Enabled`.

The PR does that and reworks the tests in mod.rs to use async backing.
It's a preparation for
https://github.com/paritytech/polkadot-sdk/issues/5079

---------
