+++ polkadot/node/core/backing/src/lib.rs
@@ -98,13 +98,10 @@ use polkadot_node_subsystem::{
 };
 use polkadot_node_subsystem_util::{
 	self as util,
-	backing_implicit_view::{FetchError as ImplicitViewFetchError, View as ImplicitView},
-	request_from_runtime, request_session_executor_params, request_session_index_for_child,
-	request_validator_groups, request_validators,
-	runtime::{
-		self, fetch_claim_queue, prospective_parachains_mode, request_min_backing_votes,
-		ClaimQueueSnapshot, ProspectiveParachainsMode,
-	},
+	backing_implicit_view::View as ImplicitView,
+	request_claim_queue, request_disabled_validators, request_session_executor_params,
+	request_session_index_for_child, request_validator_groups, request_validators,
+	runtime::{self, request_min_backing_votes, ClaimQueueSnapshot},
 	Validator,
 };
 use polkadot_parachain_primitives::primitives::IsSystem;
@@ -112,7 +109,7 @@ use polkadot_primitives::{
 	node_features::FeatureIndex,
 	vstaging::{
 		BackedCandidate, CandidateReceiptV2 as CandidateReceipt,
-		CommittedCandidateReceiptV2 as CommittedCandidateReceipt, CoreState,
+		CommittedCandidateReceiptV2 as CommittedCandidateReceipt,
 	},
 	CandidateCommitments, CandidateHash, CoreIndex, ExecutorParams, GroupIndex, GroupRotationInfo,
 	Hash, Id as ParaId, IndexedVec, NodeFeatures, PersistedValidationData, SessionIndex,
@@ -125,10 +122,10 @@ use polkadot_statement_table::{
 		SignedStatement as TableSignedStatement, Statement as TableStatement,
 		Summary as TableSummary,
 	},
-	Config as TableConfig, Context as TableContextTrait, Table,
+	Context as TableContextTrait, Table,
 };
 use sp_keystore::KeystorePtr;
-use util::runtime::{get_disabled_validators_with_fallback, request_node_features};
+use util::runtime::request_node_features;
 
 mod error;
 
@@ -215,7 +212,6 @@ where
 }
 
 struct PerRelayParentState {
-	prospective_parachains_mode: ProspectiveParachainsMode,
 	/// The hash of the relay parent on top of which this job is doing it's work.
 	parent: Hash,
 	/// The node features.
@@ -256,43 +252,6 @@ struct PerCandidateState {
 	persisted_validation_data: PersistedValidationData,
 	seconded_locally: bool,
 	relay_parent: Hash,
-}
-
-enum ActiveLeafState {
-	// If prospective-parachains is disabled, one validator may only back one candidate per
-	// paraid.
-	ProspectiveParachainsDisabled { seconded: HashSet<ParaId> },
-	ProspectiveParachainsEnabled { max_candidate_depth: usize, allowed_ancestry_len: usize },
-}
-
-impl ActiveLeafState {
-	fn new(mode: ProspectiveParachainsMode) -> Self {
-		match mode {
-			ProspectiveParachainsMode::Disabled =>
-				Self::ProspectiveParachainsDisabled { seconded: HashSet::new() },
-			ProspectiveParachainsMode::Enabled { max_candidate_depth, allowed_ancestry_len } =>
-				Self::ProspectiveParachainsEnabled { max_candidate_depth, allowed_ancestry_len },
-		}
-	}
-
-	fn add_seconded_candidate(&mut self, para_id: ParaId) {
-		if let Self::ProspectiveParachainsDisabled { seconded } = self {
-			seconded.insert(para_id);
-		}
-	}
-}
-
-impl From<&ActiveLeafState> for ProspectiveParachainsMode {
-	fn from(state: &ActiveLeafState) -> Self {
-		match *state {
-			ActiveLeafState::ProspectiveParachainsDisabled { .. } =>
-				ProspectiveParachainsMode::Disabled,
-			ActiveLeafState::ProspectiveParachainsEnabled {
-				max_candidate_depth,
-				allowed_ancestry_len,
-			} => ProspectiveParachainsMode::Enabled { max_candidate_depth, allowed_ancestry_len },
-		}
-	}
 }
 
 /// A cache for storing data per-session to reduce repeated
@@ -470,27 +429,9 @@ impl PerSessionCache {
 /// The state of the subsystem.
 struct State {
 	/// The utility for managing the implicit and explicit views in a consistent way.
-	///
-	/// We only feed leaves which have prospective parachains enabled to this view.
 	implicit_view: ImplicitView,
-	/// State tracked for all active leaves, whether or not they have prospective parachains
-	/// enabled.
-	per_leaf: HashMap<Hash, ActiveLeafState>,
 	/// State tracked for all relay-parents backing work is ongoing for. This includes
 	/// all active leaves.
-	///
-	/// relay-parents fall into one of 3 categories.
-	///   1. active leaves which do support prospective parachains
-	///   2. active leaves which do not support prospective parachains
-	///   3. relay-chain blocks which are ancestors of an active leaf and do support prospective
-	///      parachains.
-	///
-	/// Relay-chain blocks which don't support prospective parachains are
-	/// never included in the fragment chains of active leaves which do.
-	///
-	/// While it would be technically possible to support such leaves in
-	/// fragment chains, it only benefits the transition period when asynchronous
-	/// backing is being enabled and complicates code.
 	per_relay_parent: HashMap<Hash, PerRelayParentState>,
 	/// State tracked for all candidates relevant to the implicit view.
 	///
@@ -514,7 +455,6 @@ impl State {
 	) -> Self {
 		State {
 			implicit_view: ImplicitView::default(),
-			per_leaf: HashMap::default(),
 			per_relay_parent: HashMap::default(),
 			per_candidate: HashMap::new(),
 			per_session_cache: PerSessionCache::default(),
@@ -1022,86 +962,41 @@ async fn handle_active_leaves_update<Context>(
 	update: ActiveLeavesUpdate,
 	state: &mut State,
 ) -> Result<(), Error> {
-	enum LeafHasProspectiveParachains {
-		Enabled(Result<ProspectiveParachainsMode, ImplicitViewFetchError>),
-		Disabled,
-	}
-
 	// Activate in implicit view before deactivate, per the docs
 	// on ImplicitView, this is more efficient.
 	let res = if let Some(leaf) = update.activated {
-		// Only activate in implicit view if prospective
-		// parachains are enabled.
-		let mode = prospective_parachains_mode(ctx.sender(), leaf.hash).await?;
-
 		let leaf_hash = leaf.hash;
-		Some((
-			leaf,
-			match mode {
-				ProspectiveParachainsMode::Disabled => LeafHasProspectiveParachains::Disabled,
-				ProspectiveParachainsMode::Enabled { .. } => LeafHasProspectiveParachains::Enabled(
-					state.implicit_view.activate_leaf(ctx.sender(), leaf_hash).await.map(|_| mode),
-				),
-			},
-		))
+		Some((leaf, state.implicit_view.activate_leaf(ctx.sender(), leaf_hash).await.map(|_| ())))
 	} else {
 		None
 	};
 
 	for deactivated in update.deactivated {
-		state.per_leaf.remove(&deactivated);
 		state.implicit_view.deactivate_leaf(deactivated);
 	}
 
 	// clean up `per_relay_parent` according to ancestry
 	// of leaves. we do this so we can clean up candidates right after
 	// as a result.
-	//
-	// when prospective parachains are disabled, the implicit view is empty,
-	// which means we'll clean up everything that's not a leaf - the expected behavior
-	// for pre-asynchronous backing.
 	{
-		let remaining: HashSet<_> = state
-			.per_leaf
-			.keys()
-			.chain(state.implicit_view.all_allowed_relay_parents())
-			.collect();
+		let remaining: HashSet<_> = state.implicit_view.all_allowed_relay_parents().collect();
 
 		state.per_relay_parent.retain(|r, _| remaining.contains(&r));
 	}
 
 	// clean up `per_candidate` according to which relay-parents
 	// are known.
-	//
-	// when prospective parachains are disabled, we clean up all candidates
-	// because we've cleaned up all relay parents. this is correct.
 	state
 		.per_candidate
 		.retain(|_, pc| state.per_relay_parent.contains_key(&pc.relay_parent));
 
 	// Get relay parents which might be fresh but might be known already
 	// that are explicit or implicit from the new active leaf.
-	let (fresh_relay_parents, leaf_mode) = match res {
+	let fresh_relay_parents = match res {
 		None => return Ok(()),
-		Some((leaf, LeafHasProspectiveParachains::Disabled)) => {
-			// defensive in this case - for enabled, this manifests as an error.
-			if state.per_leaf.contains_key(&leaf.hash) {
-				return Ok(())
-			}
-
-			state
-				.per_leaf
-				.insert(leaf.hash, ActiveLeafState::new(ProspectiveParachainsMode::Disabled));
-
-			(vec![leaf.hash], ProspectiveParachainsMode::Disabled)
-		},
-		Some((leaf, LeafHasProspectiveParachains::Enabled(Ok(prospective_parachains_mode)))) => {
+		Some((leaf, Ok(_))) => {
 			let fresh_relay_parents =
 				state.implicit_view.known_allowed_relay_parents_under(&leaf.hash, None);
-
-			let active_leaf_state = ActiveLeafState::new(prospective_parachains_mode);
-
-			state.per_leaf.insert(leaf.hash, active_leaf_state);
 
 			let fresh_relay_parent = match fresh_relay_parents {
 				Some(f) => f.to_vec(),
@@ -1115,9 +1010,9 @@ async fn handle_active_leaves_update<Context>(
 					vec![leaf.hash]
 				},
 			};
-			(fresh_relay_parent, prospective_parachains_mode)
+			fresh_relay_parent
 		},
-		Some((leaf, LeafHasProspectiveParachains::Enabled(Err(e)))) => {
+		Some((leaf, Err(e))) => {
 			gum::debug!(
 				target: LOG_TARGET,
 				leaf_hash = ?leaf.hash,
@@ -1135,18 +1030,6 @@ async fn handle_active_leaves_update<Context>(
 			continue
 		}
 
-		let mode = match state.per_leaf.get(&maybe_new) {
-			None => {
-				// If the relay-parent isn't a leaf itself,
-				// then it is guaranteed by the prospective parachains
-				// subsystem that it is an ancestor of a leaf which
-				// has prospective parachains enabled and that the
-				// block itself did.
-				leaf_mode
-			},
-			Some(l) => l.into(),
-		};
-
 		// construct a `PerRelayParent` from the runtime API
 		// and insert it.
 		let per = construct_per_relay_parent_state(
@@ -1154,7 +1037,6 @@ async fn handle_active_leaves_update<Context>(
 			maybe_new,
 			&state.keystore,
 			&mut state.per_session_cache,
-			mode,
 		)
 		.await?;
 
@@ -1254,17 +1136,14 @@ async fn construct_per_relay_parent_state<Context>(
 	relay_parent: Hash,
 	keystore: &KeystorePtr,
 	per_session_cache: &mut PerSessionCache,
-	mode: ProspectiveParachainsMode,
 ) -> Result<Option<PerRelayParentState>, Error> {
 	let parent = relay_parent;
 
-	let (session_index, groups, cores) = futures::try_join!(
+	let (session_index, groups, claim_queue, disabled_validators) = futures::try_join!(
 		request_session_index_for_child(parent, ctx.sender()).await,
 		request_validator_groups(parent, ctx.sender()).await,
-		request_from_runtime(parent, ctx.sender(), |tx| {
-			RuntimeApiRequest::AvailabilityCores(tx)
-		},)
-		.await,
+		request_claim_queue(parent, ctx.sender()).await,
+		request_disabled_validators(parent, ctx.sender()).await,
 	)
 	.map_err(Error::JoinMultiple)?;
 
@@ -1290,23 +1169,13 @@ async fn construct_per_relay_parent_state<Context>(
 	gum::debug!(target: LOG_TARGET, inject_core_index, ?parent, "New state");
 
 	let (validator_groups, group_rotation_info) = try_runtime_api!(groups);
-	let cores = try_runtime_api!(cores);
 
 	let minimum_backing_votes = per_session_cache
 		.minimum_backing_votes(session_index, parent, ctx.sender())
 		.await;
 	let minimum_backing_votes = try_runtime_api!(minimum_backing_votes);
-
-	// TODO: https://github.com/paritytech/polkadot-sdk/issues/1940
-	// Once runtime ver `DISABLED_VALIDATORS_RUNTIME_REQUIREMENT` is released remove this call to
-	// `get_disabled_validators_with_fallback`, add `request_disabled_validators` call to the
-	// `try_join!` above and use `try_runtime_api!` to get `disabled_validators`
-	let disabled_validators =
-		get_disabled_validators_with_fallback(ctx.sender(), parent).await.map_err(|e| {
-			Error::UtilError(TryFrom::try_from(e).expect("the conversion is infallible; qed"))
-		})?;
-
-	let maybe_claim_queue = try_runtime_api!(fetch_claim_queue(ctx.sender(), parent).await);
+	let claim_queue = try_runtime_api!(claim_queue);
+	let disabled_validators = try_runtime_api!(disabled_validators);
 
 	let signing_context = SigningContext { parent_hash: parent, session_index };
 	let validator = match Validator::construct(
@@ -1328,33 +1197,15 @@ async fn construct_per_relay_parent_state<Context>(
 		},
 	};
 
-	let n_cores = cores.len();
+	let n_cores = validator_groups.len();
 
 	let mut groups = HashMap::<CoreIndex, Vec<ValidatorIndex>>::new();
 	let mut assigned_core = None;
 
-	let has_claim_queue = maybe_claim_queue.is_some();
-	let mut claim_queue = maybe_claim_queue.unwrap_or_default().0;
-
-	for (idx, core) in cores.iter().enumerate() {
+	for idx in 0..n_cores {
 		let core_index = CoreIndex(idx as _);
 
-		if !has_claim_queue {
-			match core {
-				CoreState::Scheduled(scheduled) =>
-					claim_queue.insert(core_index, [scheduled.para_id].into_iter().collect()),
-				CoreState::Occupied(occupied) if mode.is_enabled() => {
-					// Async backing makes it legal to build on top of
-					// occupied core.
-					if let Some(next) = &occupied.next_up_on_available {
-						claim_queue.insert(core_index, [next.para_id].into_iter().collect())
-					} else {
-						continue
-					}
-				},
-				_ => continue,
-			};
-		} else if !claim_queue.contains_key(&core_index) {
+		if !claim_queue.contains_key(&core_index) {
 			continue
 		}
 
@@ -1373,28 +1224,21 @@ async fn construct_per_relay_parent_state<Context>(
 
 	let table_context =
 		TableContext { validator, groups, validators: validators.to_vec(), disabled_validators };
-	let table_config = TableConfig {
-		allow_multiple_seconded: match mode {
-			ProspectiveParachainsMode::Enabled { .. } => true,
-			ProspectiveParachainsMode::Disabled => false,
-		},
-	};
 
 	Ok(Some(PerRelayParentState {
-		prospective_parachains_mode: mode,
 		parent,
 		node_features,
 		executor_params,
 		assigned_core,
 		backed: HashSet::new(),
-		table: Table::new(table_config),
+		table: Table::new(),
 		table_context,
 		issued_statements: HashSet::new(),
 		awaiting_validation: HashSet::new(),
 		fallbacks: HashMap::new(),
 		minimum_backing_votes,
 		inject_core_index,
-		n_cores: cores.len() as u32,
+		n_cores: validator_groups.len() as u32,
 		claim_queue: ClaimQueueSnapshot::from(claim_queue),
 		validator_to_group,
 		group_rotation_info,
@@ -1412,7 +1256,6 @@ enum SecondingAllowed {
 #[overseer::contextbounds(CandidateBacking, prefix = self::overseer)]
 async fn seconding_sanity_check<Context>(
 	ctx: &mut Context,
-	active_leaves: &HashMap<Hash, ActiveLeafState>,
 	implicit_view: &ImplicitView,
 	hypothetical_candidate: HypotheticalCandidate,
 ) -> SecondingAllowed {
@@ -1423,49 +1266,36 @@ async fn seconding_sanity_check<Context>(
 	let candidate_relay_parent = hypothetical_candidate.relay_parent();
 	let candidate_hash = hypothetical_candidate.candidate_hash();
 
-	for (head, leaf_state) in active_leaves {
-		if ProspectiveParachainsMode::from(leaf_state).is_enabled() {
-			// Check that the candidate relay parent is allowed for para, skip the
-			// leaf otherwise.
-			let allowed_parents_for_para =
-				implicit_view.known_allowed_relay_parents_under(head, Some(candidate_para));
-			if !allowed_parents_for_para.unwrap_or_default().contains(&candidate_relay_parent) {
-				continue
-			}
-
-			let (tx, rx) = oneshot::channel();
-			ctx.send_message(ProspectiveParachainsMessage::GetHypotheticalMembership(
-				HypotheticalMembershipRequest {
-					candidates: vec![hypothetical_candidate.clone()],
-					fragment_chain_relay_parent: Some(*head),
-				},
-				tx,
-			))
-			.await;
-			let response = rx.map_ok(move |candidate_memberships| {
-				let is_member_or_potential = candidate_memberships
-					.into_iter()
-					.find_map(|(candidate, leaves)| {
-						(candidate.candidate_hash() == candidate_hash).then_some(leaves)
-					})
-					.and_then(|leaves| leaves.into_iter().find(|leaf| leaf == head))
-					.is_some();
-
-				(is_member_or_potential, head)
-			});
-			responses.push_back(response.boxed());
-		} else {
-			if *head == candidate_relay_parent {
-				if let ActiveLeafState::ProspectiveParachainsDisabled { seconded } = leaf_state {
-					if seconded.contains(&candidate_para) {
-						// The leaf is already occupied. For non-prospective parachains, we only
-						// second one candidate.
-						return SecondingAllowed::No
-					}
-				}
-				responses.push_back(futures::future::ok((true, head)).boxed());
-			}
+	for head in implicit_view.leaves() {
+		// Check that the candidate relay parent is allowed for para, skip the
+		// leaf otherwise.
+		let allowed_parents_for_para =
+			implicit_view.known_allowed_relay_parents_under(head, Some(candidate_para));
+		if !allowed_parents_for_para.unwrap_or_default().contains(&candidate_relay_parent) {
+			continue
 		}
+
+		let (tx, rx) = oneshot::channel();
+		ctx.send_message(ProspectiveParachainsMessage::GetHypotheticalMembership(
+			HypotheticalMembershipRequest {
+				candidates: vec![hypothetical_candidate.clone()],
+				fragment_chain_relay_parent: Some(*head),
+			},
+			tx,
+		))
+		.await;
+		let response = rx.map_ok(move |candidate_memberships| {
+			let is_member_or_potential = candidate_memberships
+				.into_iter()
+				.find_map(|(candidate, leaves)| {
+					(candidate.candidate_hash() == candidate_hash).then_some(leaves)
+				})
+				.and_then(|leaves| leaves.into_iter().find(|leaf| leaf == head))
+				.is_some();
+
+			(is_member_or_potential, head)
+		});
+		responses.push_back(response.boxed());
 	}
 
 	if responses.is_empty() {
@@ -1514,11 +1344,7 @@ async fn handle_can_second_request<Context>(
 	tx: oneshot::Sender<bool>,
 ) {
 	let relay_parent = request.candidate_relay_parent;
-	let response = if state
-		.per_relay_parent
-		.get(&relay_parent)
-		.map_or(false, |pr_state| pr_state.prospective_parachains_mode.is_enabled())
-	{
+	let response = if state.per_relay_parent.get(&relay_parent).is_some() {
 		let hypothetical_candidate = HypotheticalCandidate::Incomplete {
 			candidate_hash: request.candidate_hash,
 			candidate_para: request.candidate_para_id,
@@ -1526,13 +1352,8 @@ async fn handle_can_second_request<Context>(
 			candidate_relay_parent: relay_parent,
 		};
 
-		let result = seconding_sanity_check(
-			ctx,
-			&state.per_leaf,
-			&state.implicit_view,
-			hypothetical_candidate,
-		)
-		.await;
+		let result =
+			seconding_sanity_check(ctx, &state.implicit_view, hypothetical_candidate).await;
 
 		match result {
 			SecondingAllowed::No => false,
@@ -1585,16 +1406,14 @@ async fn handle_validated_candidate_command<Context>(
 						// sanity check that we're allowed to second the candidate
 						// and that it doesn't conflict with other candidates we've
 						// seconded.
-						let hypothetical_membership = match seconding_sanity_check(
+						if let SecondingAllowed::No = seconding_sanity_check(
 							ctx,
-							&state.per_leaf,
 							&state.implicit_view,
 							hypothetical_candidate,
 						)
 						.await
 						{
-							SecondingAllowed::No => return Ok(()),
-							SecondingAllowed::Yes(membership) => membership,
+							return Ok(())
 						};
 
 						let statement =
@@ -1642,24 +1461,6 @@ async fn handle_validated_candidate_command<Context>(
 									);
 								},
 								Some(p) => p.seconded_locally = true,
-							}
-
-							// record seconded candidates for non-prospective-parachains mode.
-							for leaf in hypothetical_membership {
-								let leaf_data = match state.per_leaf.get_mut(&leaf) {
-									None => {
-										gum::warn!(
-											target: LOG_TARGET,
-											leaf_hash = ?leaf,
-											"Missing `per_leaf` for known active leaf."
-										);
-
-										continue
-									},
-									Some(d) => d,
-								};
-
-								leaf_data.add_seconded_candidate(candidate.descriptor().para_id());
 							}
 
 							rp_state.issued_statements.insert(candidate_hash);
@@ -1765,13 +1566,11 @@ fn sign_statement(
 
 /// Import a statement into the statement table and return the summary of the import.
 ///
-/// This will fail with `Error::RejectedByProspectiveParachains` if the message type
-/// is seconded, the candidate is fresh,
-/// and any of the following are true:
+/// This will fail with `Error::RejectedByProspectiveParachains` if the message type is seconded,
+/// the candidate is fresh, and any of the following are true:
 /// 1. There is no `PersistedValidationData` attached.
-/// 2. Prospective parachains are enabled for the relay parent and the prospective parachains
-///    subsystem returned an empty `HypotheticalMembership` i.e. did not recognize the candidate as
-///    being applicable to any of the active leaves.
+/// 2. Prospective parachains subsystem returned an empty `HypotheticalMembership` i.e. did not
+///    recognize the candidate as being applicable to any of the active leaves.
 #[overseer::contextbounds(CandidateBacking, prefix = self::overseer)]
 async fn import_statement<Context>(
 	ctx: &mut Context,
@@ -1792,8 +1591,7 @@ async fn import_statement<Context>(
 	// If this is a new candidate (statement is 'seconded' and candidate is unknown),
 	// we need to create an entry in the `PerCandidateState` map.
 	//
-	// If the relay parent supports prospective parachains, we also need
-	// to inform the prospective parachains subsystem of the seconded candidate.
+	// We also need to inform the prospective parachains subsystem of the seconded candidate.
 	// If `ProspectiveParachainsMessage::Second` fails, then we return
 	// Error::RejectedByProspectiveParachains.
 	//
@@ -1804,30 +1602,28 @@ async fn import_statement<Context>(
 	// our active leaves.
 	if let StatementWithPVD::Seconded(candidate, pvd) = statement.payload() {
 		if !per_candidate.contains_key(&candidate_hash) {
-			if rp_state.prospective_parachains_mode.is_enabled() {
-				let (tx, rx) = oneshot::channel();
-				ctx.send_message(ProspectiveParachainsMessage::IntroduceSecondedCandidate(
-					IntroduceSecondedCandidateRequest {
-						candidate_para: candidate.descriptor.para_id(),
-						candidate_receipt: candidate.clone(),
-						persisted_validation_data: pvd.clone(),
-					},
-					tx,
-				))
-				.await;
+			let (tx, rx) = oneshot::channel();
+			ctx.send_message(ProspectiveParachainsMessage::IntroduceSecondedCandidate(
+				IntroduceSecondedCandidateRequest {
+					candidate_para: candidate.descriptor.para_id(),
+					candidate_receipt: candidate.clone(),
+					persisted_validation_data: pvd.clone(),
+				},
+				tx,
+			))
+			.await;
 
-				match rx.await {
-					Err(oneshot::Canceled) => {
-						gum::warn!(
-							target: LOG_TARGET,
-							"Could not reach the Prospective Parachains subsystem."
-						);
+			match rx.await {
+				Err(oneshot::Canceled) => {
+					gum::warn!(
+						target: LOG_TARGET,
+						"Could not reach the Prospective Parachains subsystem."
+					);
 
-						return Err(Error::RejectedByProspectiveParachains)
-					},
-					Ok(false) => return Err(Error::RejectedByProspectiveParachains),
-					Ok(true) => {},
-				}
+					return Err(Error::RejectedByProspectiveParachains)
+				},
+				Ok(false) => return Err(Error::RejectedByProspectiveParachains),
+				Ok(true) => {},
 			}
 
 			// Only save the candidate if it was approved by prospective parachains.
@@ -1890,28 +1686,15 @@ async fn post_import_statement_actions<Context>(
 					"Candidate backed",
 				);
 
-				if rp_state.prospective_parachains_mode.is_enabled() {
-					// Inform the prospective parachains subsystem
-					// that the candidate is now backed.
-					ctx.send_message(ProspectiveParachainsMessage::CandidateBacked(
-						para_id,
-						candidate_hash,
-					))
-					.await;
-					// Notify statement distribution of backed candidate.
-					ctx.send_message(StatementDistributionMessage::Backed(candidate_hash)).await;
-				} else {
-					// The provisioner waits on candidate-backing, which means
-					// that we need to send unbounded messages to avoid cycles.
-					//
-					// Backed candidates are bounded by the number of validators,
-					// parachains, and the block production rate of the relay chain.
-					let message = ProvisionerMessage::ProvisionableData(
-						rp_state.parent,
-						ProvisionableData::BackedCandidate(backed.receipt()),
-					);
-					ctx.send_unbounded_message(message);
-				}
+				// Inform the prospective parachains subsystem
+				// that the candidate is now backed.
+				ctx.send_message(ProspectiveParachainsMessage::CandidateBacked(
+					para_id,
+					candidate_hash,
+				))
+				.await;
+				// Notify statement distribution of backed candidate.
+				ctx.send_message(StatementDistributionMessage::Backed(candidate_hash)).await;
 			} else {
 				gum::debug!(target: LOG_TARGET, ?candidate_hash, "Cannot get BackedCandidate");
 			}
