+++ polkadot/statement-table/src/generic.rs
@@ -62,14 +62,6 @@ pub trait Context {
 	fn get_group_size(&self, group: &Self::GroupId) -> Option<usize>;
 }
 
-/// Table configuration.
-pub struct Config {
-	/// When this is true, the table will allow multiple seconded candidates
-	/// per authority. This flag means that higher-level code is responsible for
-	/// bounding the number of candidates.
-	pub allow_multiple_seconded: bool,
-}
-
 /// Statements circulated among peers.
 #[derive(PartialEq, Eq, Debug, Clone, Encode, Decode)]
 pub enum Statement<Candidate, Digest> {
@@ -143,15 +135,6 @@ impl<Candidate, Digest, Signature> DoubleSign<Candidate, Digest, Signature> {
 	}
 }
 
-/// Misbehavior: declaring multiple candidates.
-#[derive(PartialEq, Eq, Debug, Clone)]
-pub struct MultipleCandidates<Candidate, Signature> {
-	/// The first candidate seen.
-	pub first: (Candidate, Signature),
-	/// The second candidate seen.
-	pub second: (Candidate, Signature),
-}
-
 /// Misbehavior: submitted statement for wrong group.
 #[derive(PartialEq, Eq, Debug, Clone)]
 pub struct UnauthorizedStatement<Candidate, Digest, AuthorityId, Signature> {
@@ -165,8 +148,6 @@ pub struct UnauthorizedStatement<Candidate, Digest, AuthorityId, Signature> {
 pub enum Misbehavior<Candidate, Digest, AuthorityId, Signature> {
 	/// Voted invalid and valid on validity.
 	ValidityDoubleVote(ValidityDoubleVote<Candidate, Digest, Signature>),
-	/// Submitted multiple candidates.
-	MultipleCandidates(MultipleCandidates<Candidate, Signature>),
 	/// Submitted a message that was unauthorized.
 	UnauthorizedStatement(UnauthorizedStatement<Candidate, Digest, AuthorityId, Signature>),
 	/// Submitted two valid signatures for the same message.
@@ -300,17 +281,14 @@ pub struct Table<Ctx: Context> {
 	authority_data: HashMap<Ctx::AuthorityId, AuthorityData<Ctx>>,
 	detected_misbehavior: HashMap<Ctx::AuthorityId, Vec<MisbehaviorFor<Ctx>>>,
 	candidate_votes: HashMap<Ctx::Digest, CandidateData<Ctx>>,
-	config: Config,
 }
 
 impl<Ctx: Context> Table<Ctx> {
-	/// Create a new `Table` from a `Config`.
-	pub fn new(config: Config) -> Self {
+	pub fn new() -> Self {
 		Table {
 			authority_data: HashMap::default(),
 			detected_misbehavior: HashMap::default(),
 			candidate_votes: HashMap::default(),
-			config,
 		}
 	}
 
@@ -408,33 +386,7 @@ impl<Ctx: Context> Table<Ctx> {
 				// if digest is different, fetch candidate and
 				// note misbehavior.
 				let existing = occ.get_mut();
-
-				if !self.config.allow_multiple_seconded && existing.proposals.len() == 1 {
-					let (old_digest, old_sig) = &existing.proposals[0];
-
-					if old_digest != &digest {
-						const EXISTENCE_PROOF: &str =
-							"when proposal first received from authority, candidate \
-							votes entry is created. proposal here is `Some`, therefore \
-							candidate votes entry exists; qed";
-
-						let old_candidate = self
-							.candidate_votes
-							.get(old_digest)
-							.expect(EXISTENCE_PROOF)
-							.candidate
-							.clone();
-
-						return Err(Misbehavior::MultipleCandidates(MultipleCandidates {
-							first: (old_candidate, old_sig.clone()),
-							second: (candidate, signature.clone()),
-						}))
-					}
-
-					false
-				} else if self.config.allow_multiple_seconded &&
-					existing.proposals.iter().any(|(ref od, _)| od == &digest)
-				{
+				if existing.proposals.iter().any(|(ref od, _)| od == &digest) {
 					false
 				} else {
 					existing.proposals.push((digest.clone(), signature.clone()));
@@ -591,14 +543,6 @@ mod tests {
 	use super::*;
 	use std::collections::HashMap;
 
-	fn create_single_seconded<Candidate: Context>() -> Table<Candidate> {
-		Table::new(Config { allow_multiple_seconded: false })
-	}
-
-	fn create_many_seconded<Candidate: Context>() -> Table<Candidate> {
-		Table::new(Config { allow_multiple_seconded: true })
-	}
-
 	#[derive(Debug, Copy, Clone, Hash, PartialEq, Eq)]
 	struct AuthorityId(usize);
 
@@ -647,42 +591,6 @@ mod tests {
 	}
 
 	#[test]
-	fn submitting_two_candidates_can_be_misbehavior() {
-		let context = TestContext {
-			authorities: {
-				let mut map = HashMap::new();
-				map.insert(AuthorityId(1), GroupId(2));
-				map
-			},
-		};
-
-		let mut table = create_single_seconded();
-		let statement_a = SignedStatement {
-			statement: Statement::Seconded(Candidate(2, 100)),
-			signature: Signature(1),
-			sender: AuthorityId(1),
-		};
-
-		let statement_b = SignedStatement {
-			statement: Statement::Seconded(Candidate(2, 999)),
-			signature: Signature(1),
-			sender: AuthorityId(1),
-		};
-
-		table.import_statement(&context, GroupId(2), statement_a);
-		assert!(!table.detected_misbehavior.contains_key(&AuthorityId(1)));
-
-		table.import_statement(&context, GroupId(2), statement_b);
-		assert_eq!(
-			table.detected_misbehavior[&AuthorityId(1)][0],
-			Misbehavior::MultipleCandidates(MultipleCandidates {
-				first: (Candidate(2, 100), Signature(1)),
-				second: (Candidate(2, 999), Signature(1)),
-			})
-		);
-	}
-
-	#[test]
 	fn submitting_two_candidates_can_be_allowed() {
 		let context = TestContext {
 			authorities: {
@@ -692,7 +600,7 @@ mod tests {
 			},
 		};
 
-		let mut table = create_many_seconded();
+		let mut table = Table::new();
 		let statement_a = SignedStatement {
 			statement: Statement::Seconded(Candidate(2, 100)),
 			signature: Signature(1),
@@ -722,7 +630,7 @@ mod tests {
 			},
 		};
 
-		let mut table = create_single_seconded();
+		let mut table = Table::new();
 		let statement = SignedStatement {
 			statement: Statement::Seconded(Candidate(2, 100)),
 			signature: Signature(1),
@@ -754,7 +662,7 @@ mod tests {
 			},
 		};
 
-		let mut table = create_single_seconded();
+		let mut table = Table::new();
 
 		let candidate_a = SignedStatement {
 			statement: Statement::Seconded(Candidate(2, 100)),
@@ -798,7 +706,7 @@ mod tests {
 			},
 		};
 
-		let mut table = create_single_seconded();
+		let mut table = Table::new();
 		let statement = SignedStatement {
 			statement: Statement::Seconded(Candidate(2, 100)),
 			signature: Signature(1),
@@ -828,7 +736,7 @@ mod tests {
 			},
 		};
 
-		let mut table = create_single_seconded();
+		let mut table = Table::new();
 		let statement = SignedStatement {
 			statement: Statement::Seconded(Candidate(2, 100)),
 			signature: Signature(1),
@@ -896,7 +804,7 @@ mod tests {
 		};
 
 		// have 2/3 validity guarantors note validity.
-		let mut table = create_single_seconded();
+		let mut table = Table::new();
 		let statement = SignedStatement {
 			statement: Statement::Seconded(Candidate(2, 100)),
 			signature: Signature(1),
@@ -930,7 +838,7 @@ mod tests {
 			},
 		};
 
-		let mut table = create_single_seconded();
+		let mut table = Table::new();
 		let statement = SignedStatement {
 			statement: Statement::Seconded(Candidate(2, 100)),
 			signature: Signature(1),
@@ -957,7 +865,7 @@ mod tests {
 			},
 		};
 
-		let mut table = create_single_seconded();
+		let mut table = Table::new();
 		let statement = SignedStatement {
 			statement: Statement::Seconded(Candidate(2, 100)),
 			signature: Signature(1),
