+++ polkadot/statement-table/src/lib.rs
@@ -29,7 +29,7 @@
 
 pub mod generic;
 
-pub use generic::{Config, Context, Table};
+pub use generic::{Context, Table};
 
 /// Concrete instantiations suitable for v2 primitives.
 pub mod v2 {
